Add support for OpenSSL 3
