+++ src/crypto.h
@@ -1,19 +1,10 @@
 #pragma once
 
-#include <openssl/hmac.h>
 #include <sframe/sframe.h>
 
 #include <array>
 
 namespace sframe {
-
-///
-/// Scoped pointers for OpenSSL objects
-///
-
-using scoped_evp_ctx =
-  std::unique_ptr<EVP_CIPHER_CTX, decltype(&EVP_CIPHER_CTX_free)>;
-using scoped_hmac_ctx = std::unique_ptr<HMAC_CTX, decltype(&HMAC_CTX_free)>;
 
 ///
 /// Information about algorithms
@@ -24,21 +15,15 @@ cipher_digest_size(CipherSuite suite);
 size_t
 cipher_key_size(CipherSuite suite);
 size_t
+cipher_enc_key_size(CipherSuite suite);
+size_t
 cipher_nonce_size(CipherSuite suite);
+size_t
+cipher_overhead(CipherSuite suite);
 
 ///
 /// HMAC and HKDF
 ///
-
-struct HMAC
-{
-  HMAC(CipherSuite suite, input_bytes key);
-  void write(input_bytes data);
-  input_bytes digest();
-
-  scoped_hmac_ctx ctx;
-  std::array<uint8_t, EVP_MAX_MD_SIZE> md;
-};
 
 bytes
 hkdf_extract(CipherSuite suite, const bytes& salt, const bytes& ikm);
