+++ test/sframe.cpp
@@ -10,15 +10,6 @@
 
 using namespace sframe;
 
-static void
-ensure_fips_if_required()
-{
-  const auto* require = std::getenv("REQUIRE_FIPS");
-  if (require && FIPS_mode() == 0) {
-    REQUIRE(FIPS_mode_set(1) == 1);
-  }
-}
-
 static bytes
 from_hex(const std::string& hex)
 {
@@ -45,8 +36,6 @@ to_bytes(const T& range)
 
 TEST_CASE("SFrame Known-Answer")
 {
-  ensure_fips_if_required();
-
   struct KnownAnswerTest
   {
     bytes key;
@@ -143,8 +132,6 @@ TEST_CASE("SFrame Known-Answer")
 
 TEST_CASE("SFrame Round-Trip")
 {
-  ensure_fips_if_required();
-
   const auto rounds = 1 << 9;
   const auto kid = KeyID(0x42);
   const auto plaintext = from_hex("00010203");
@@ -183,8 +170,6 @@ TEST_CASE("SFrame Round-Trip")
 
 TEST_CASE("MLS Known-Answer")
 {
-  ensure_fips_if_required();
-
   struct KnownAnswerTest
   {
     using Epoch = std::vector<bytes>;
@@ -312,8 +297,6 @@ TEST_CASE("MLS Known-Answer")
 
 TEST_CASE("MLS Round-Trip")
 {
-  ensure_fips_if_required();
-
   const auto epoch_bits = 2;
   const auto test_epochs = 1 << (epoch_bits + 1);
   const auto epoch_rounds = 10;
@@ -359,8 +342,6 @@ TEST_CASE("MLS Round-Trip")
 
 TEST_CASE("MLS Known-Answer with Context")
 {
-  ensure_fips_if_required();
-
   struct KnownAnswerTest
   {
     using ContextCases = std::vector<bytes>;
@@ -643,8 +624,6 @@ TEST_CASE("MLS Known-Answer with Context")
 
 TEST_CASE("MLS Round-Trip with context")
 {
-  ensure_fips_if_required();
-
   const auto epoch_bits = 4;
   const auto test_epochs = 1 << (epoch_bits + 1);
   const auto epoch_rounds = 10;
@@ -709,8 +688,6 @@ TEST_CASE("MLS Round-Trip with context")
 
 TEST_CASE("MLS Failure after Purge")
 {
-  ensure_fips_if_required();
-
   const auto suite = CipherSuite::AES_GCM_128_SHA256;
   const auto epoch_bits = 2;
   const auto plaintext = from_hex("00010203");
